requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
marshmallow==3.26.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
Werkzeug==3.1.5
prometheus-flask-exporter==0.23.0
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy
python-json-logger==2.0.7
boto3==1.34.0